% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brickchart.R
\name{brickchart}
\alias{brickchart}
\title{Brick Chart}
\usage{
brickchart(
  data,
  outcome,
  by,
  group,
  colors = list(c((viridis::viridis_pal(end = 0.9, option = "cividis", alpha =
    1))(3)[1], (viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8))(3)[1]),
    c((viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 1))(3)[2],
    (viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8))(3)[2]),
    c((viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 1))(3)[3],
    (viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8))(3)[3])),
  guide = FALSE
)
}
\arguments{
\item{data}{Data set}

\item{outcome}{Outcome expression, e.g., \code{event == TRUE}}

\item{by}{Exposure variable}

\item{group}{Optional: Grouping variable, e.g. an effect modifier.}

\item{colors}{Color list. Must be a \code{list} consisting of two-element
color code vectors with the dark and bright colors. The default
provides colors for three exposure groups; additions may be necessary.
Example: \code{list(c("darkred", "red"), c("darkblue", "lightblue"))}.}

\item{guide}{Optional: Show legend? Defaults to \code{FALSE}.}
}
\value{
ggplot. Modify further with standard ggplot functions.
}
\description{
Plot bars of proportions that consist of "bricks" showing individual
observations.
}
\examples{
data(ovarian, package = "survival")
ovarian \%>\%
  brickchart(outcome = fustat == 1,
             by = ecog.ps)
ovarian \%>\%
  brickchart(outcome = fustat == 1,
             by = ecog.ps,
             group = rx)
}
