% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrmat}
\alias{corrmat}
\title{Correlation plot}
\usage{
corrmat(
  data,
  method = "pearson",
  missing = "pairwise.complete.obs",
  reorder = TRUE,
  digits = 2,
  legendpos = c(0.15, 0.35),
  cutpoints = c(-1, 0, 1),
  colors = c(viridis_pal(option = "cividis")(10)[1], "white", viridis_pal(option =
    "cividis")(10)[8])
)
}
\arguments{
\item{data}{Required: Data frame. Only numeric variables will be used;
variables of other types will silently be dropped. To include
categorical variables, coerce factors to numeric first. See examples.}

\item{method}{Correlation method. Defaults to \code{"pearson"}.}

\item{missing}{How \code{\link[stats]{cor}} handles
missing values. Defaults to a restriction to complete
observations with all variables. See help for
\code{\link[stats]{cor}} for alternatives.}

\item{reorder}{Perform hierarchical clustering in the correlation matrix?
This will order variables by their correlation patterns with other variables.
If turned off to \code{FALSE}, the order of variables in the dataset will be retained.
Defaults to \code{TRUE}.}

\item{digits}{Decimal digits for displayed correlation coefficients. Defaults to \code{2}.}

\item{legendpos}{(x, y) coordinates of color legend. Use
\code{legendpos = "none"} to turn off the legend.
Defaults to \code{c(0.15, 0.35)}.}

\item{cutpoints}{Correlation coefficient values that have a distinct
color. Defaults to \code{c(-1, 0, 1)}.}

\item{colors}{Colors for the \code{cutpoints}. Defaults to blue (for negative),
white (no correlation), and yellow (positive correlation) on the
\code{cividis} color scale.}
}
\value{
ggplot. Can be modified with the usual ggplot commands, such as
\code{\link[ggplot2]{theme}}.
}
\description{
Performs pairwise Pearson or Spearman correlations of all
numeric variables that exist in the provided data frame.
}
\examples{
data(mtcars)

mtcars \%>\%
  select(mpg, cyl, hp, wt, qsec) \%>\%
  corrmat()

# If "cyl" was a character, it would be excluded:
mtcars \%>\%
  mutate(cyl_chr = as.character(cyl)) \%>\%
  select(mpg, cyl_chr, hp, wt, qsec) \%>\%
  corrmat()

# To retain the character variable "cyl",
# convert to factor and then make numeric:
mtcars \%>\%
  mutate(cyl_chr = as.character(cyl),
         cyl_chr = as.numeric(factor(cyl_chr))) \%>\%
  select(mpg, cyl_chr, hp, wt, qsec) \%>\%
  corrmat()
}
\seealso{
\url{http://www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization}
}
