% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtl.R
\name{estimate_rmtl}
\alias{estimate_rmtl}
\title{Estimate Restricted Mean Time Lost (RMTL) and its Difference}
\usage{
estimate_rmtl(
  data,
  exposure = NULL,
  time,
  time2 = NULL,
  event,
  event_of_interest = 1,
  weight = NULL,
  tau = NULL,
  conf.level = 0.95
)
}
\arguments{
\item{data}{Data frame (tibble).}

\item{exposure}{Optional. Name of exposure variable within the \code{data}.
If omitted, will return unstratified RMTL.}

\item{time}{Name of time variable within the \code{data}. By default, this
is the event time. If \code{time2} is also given, this is the entry time,
as per \code{\link[survival]{Surv}} standard.}

\item{time2}{Optional. Name of the variable  within the \code{data}
indicating the exit (event) time. If omitted, \code{time} is  the event
time, as per \code{\link[survival]{Surv}} standard.}

\item{event}{Name of event variable within the \code{data}. Can be integer,
factor, character, or logical. The reference level indicates censoring and
is taken as the first level of the factor, the lowest numeric value
(usually \code{0}), or \code{FALSE} for a logical variables. Other levels
are events of different types. If no competing risks, the alternative
level, e.g. \code{1}, indicates the event (e.g., death).}

\item{event_of_interest}{Optional. Indicator for which of the non-censoring
events is of main interest. Others are treated as competing. Defaults to
\code{1}, i.e., the first non-censoring event.}

\item{weight}{Optional. Weights, such as inverse-probability weights or
survey weights. The default (\code{NULL}) uses equal weights for all
observations.}

\item{tau}{Optional. Time horizon to restrict event times. By default, the
the latest time in the group with the shortest follow-up. Prefer a
user-defined, interpretable time horizon.}

\item{conf.level}{Optional. Confidence level. Defaults to \code{0.95}.}
}
\value{
List:
\itemize{
\item \code{tau} Time horizon.
\item \code{rmtl} Tibble with absolute estimates of RMTL, per exposure group if
given.
\item \code{rmtdiff} Tibble with contrasts of RMTL between exposure groups,
compared to a common reference (the first level). Empty tibble if no
exposure variable given.
\item \code{cif} Tibble with cumulative incidence function for the event of
interest:
\itemize{
\item \code{exposure} Exposure group.
\item \code{time} Event time.
\item \code{estimate} Nelsen-Aalen estimate of cumulative incidence function
with \code{se}, \code{conf.low}, and \code{conf.high}.
}
}
}
\description{
The method of Connor and Trinquart (Stat Med 2021) estimates the RMTL in the
presence of competing risks. This function a modification of their R code.
Estimation functions are identical; input and output handling have been
adapted.
}
\details{
Differences to the original function rmtl::rmtl():
\itemize{
\item Convert print() to stop() for errors.
\item Pass data as a data frame/tibble, select variables from it.
\item Allow for different variable types in the \code{event} variable.
\item Use \code{\link[survival]{Surv}} conventions for entry/exit times
(\code{time}, \code{time2}).
\item Return contrasts in restricted mean time lost as comparisons to the
reference level instead of all pairwise comparisons.
\item Add origin (time 0, cumulative incidence 0) to the returned cumulative
incidence function for proper plotting.
\item Return results as a list of tibbles. Do not print or plot.
}
}
\examples{
data(cancer, package = "survival")
cancer <- cancer \%>\% dplyr::mutate(
  status = status - 1,   # make 0/1
  sex    = factor(sex, levels = 1:2, labels = c("Men", "Women")))

result <- estimate_rmtl(
  data = cancer,
  exposure = sex,
  time = time,
  event = status,
  tau = 365.25)  # time horizon: one year
result

# Make (simple) plot
library(ggplot2)
result$cif \%>\%
  ggplot(mapping = aes(x = time, y = estimate, color = exposure)) +
  geom_step() +
  scale_x_continuous(breaks = seq(from = 0, to = 365, by = 60)) +
  labs(x = "Time since start of follow-up in days", y = "Cumulative incidence")
}
\references{
Conner SC, Trinquart L. Estimation and modeling of the restricted
mean time lost in the presence of competing risks. Stat Med 2021;40:2177–96.
https://doi.org/10.1002/sim.8896.
}
