% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{exclusion_flowchart}
\alias{exclusion_flowchart}
\title{Exclusion Flowchart using DiagrammeR}
\usage{
exclusion_flowchart(design, width = 4)
}
\arguments{
\item{design}{Tibble with the following columns:
\itemize{
\item \code{left} Text for the left box of each row.
\item \code{n_left} Count to be shown as 'n =' in the bottom
of the left box. To skip, use \code{NA_integer_}.
\item \code{right} Text for the right box of each row.
\item \code{n_right} Count to be shown as 'n =' in the bottom
of the right box. To skip, use \code{NA_integer_}.
}}

\item{width}{Minimum width for all boxes.}
}
\value{
Plot rendered using \code{\link[DiagrammeR]{grViz}}.
For a \code{design} with three rows,
the following pseudo-code would be generated: \preformatted{
  grViz(paste0("digraph flowchart {
    node [fontname = Helvetica, shape = rectangle, width = 4]
    tab1a [label = 'left[1] \\nn = ", n_left[1],  "']
    tab1b [label = 'right[1]\\nn = ", n_right[1], "']
    tab2a [label = 'left[2] \\nn = ", n_left[2],  "']
    tab2b [label = 'right[2]\\nn = ", n_right[2], "']
    tab3a [label = 'left[3] \\nn = ", n_left[3],  "']

    { rank = same; tab1a; tab1b; }
    { rank = same; tab2a; tab2b; }

    tab1a -> tab2a -> tab3a;
    tab1a -> tab1b;
    tab2a -> tab2b;
    }"))
  }
}
\description{
This function plots a flow chart using \code{\link[DiagrammeR]{grViz}}.
It is restricted to the simple case of sequential exclusions from a single
study population and not suited for CONSORT flowcharts for a
parallel-group study (i.e., a randomized-controlled trial).
The \code{left} column describes the flow of the group that ends up
being included. The \code{right} column describes exclusions for various
reasons.
}
\details{
Note \code{\link[DiagrammeR]{grViz}} does not automatically
generate line breaks. To avoid extra-wide boxes, manually
supply line breaks using \code{\\n}. See example.
}
\examples{
# Generate a flow chart for two steps of exclusions:
design <- tribble(
  ~left,               ~n_left, ~right,              ~n_right,
  "Study base",        1000,    "Not sampled",       250,
  "Study population",  750,     "Participants with\nmissing exposure data", 100,
  "Complete-case set", 650,     "",                 NA_integer_)
exclusion_flowchart(design, width = 2)
}
