% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\docType{import}
\name{expr}
\alias{expr}
\title{expr function}
\usage{
expr(expr)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:nse-defuse]{expr}}}
}}

