% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{Mode}
\alias{Mode}
\title{Most Frequent Observation}
\usage{
Mode(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Scalar: most frequent value.
}
\description{
Not to be confused with \code{mode} in lowercase,
a \code{base} function that returns the storage mode of an object.
}
\examples{
Mode(c(1, "b", "b", 0.3))
}
