% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mygt}
\alias{mygt}
\title{Turn tibble into gt Table with Custom Formatting}
\usage{
mygt(df)
}
\arguments{
\item{df}{Data frame/tibble}
}
\value{
Formatted gt table
}
\description{
Formatting includes:
\itemize{
\item Text align to top/left
\item Smaller row padding
\item No top border
\item Bold column labels
}
}
\examples{
data(mtcars)
mtcars \%>\%
  slice(1:5) \%>\%
  mygt()
}
