% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mytab}
\alias{mytab}
\title{Contingency Table}
\usage{
mytab(data, maxlevels = 30, rowsums = TRUE, colsums = TRUE)
}
\arguments{
\item{data}{Data frame (tibble).}

\item{maxlevels}{Optional. Maximum number of levels for rows and columns.
If a row or column variable has more than this number of levels,
for example because it is actually a continuous variable,
then only a warning, not a hard error, will be shown. Defaults to 30.}

\item{rowsums}{Optional. Show row sums? Defaults to TRUE.}

\item{colsums}{Optional. Show column sums? Defaults to TRUE.}
}
\value{
Tibble.
}
\description{
Tabulates X times Y. Using a data frame (tibble)
as input, the first and second columns of the data will be used
as rows and columns, respectively.
}
\examples{
data(mtcars)
mtcars \%>\%
  select(gear, carb) \%>\%
  mytab()

# Swap rows and columns, turn off totals:
mtcars \%>\%
  select(carb, gear) \%>\%
  mytab(rowsum = FALSE, colsum = FALSE)
}
