% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{pdf_safely}
\alias{pdf_safely}
\title{Safely Export as PDF Without Overwriting the File}
\usage{
pdf_safely(file = stop("'file' must be specified"), ...)
}
\arguments{
\item{file}{Path/file name to for PDF file. Required.}

\item{...}{Optional. Passed on to \code{\link[grDevices]{pdf}}.}
}
\value{
None.
}
\description{
Wraps \code{\link[grDevices]{pdf}}. If the file already exists,
it will not be overwritten. A message will be printed indicating whether the file
was successfully written or if it already existed.
}
\examples{
\dontrun{
pdf_safely(file = "graphs.pdf")
# ... code that generates plots ...
dev.off()
}
}
