% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rates}
\alias{rates}
\title{Event Counts, Person-Time, Incidence Rates, and Cumulative Incidence (Risk).}
\usage{
rates(
  data,
  time,
  event,
  conversion = 1,
  rate_digits = 1,
  risk_digits = 1,
  factor = 1000
)
}
\arguments{
\item{data}{Data frame/tibble. If grouped using \code{\link[dplyr]{group_by}},
then stratified results will be returned.}

\item{time}{Variable indicating follow-up time.}

\item{event}{Variable indicating events (event = 1, censor = 0).}

\item{conversion}{Divisor for follow-up time. Defaults to 1.
Set \code{conversion = 12} if \code{time} is in months and
rates per person-years are desired.}

\item{rate_digits}{Decimal digits to display for rates. Defaults to 1.}

\item{risk_digits}{Decimal digits to display for
cumulative incidence (risk). Defaults to 1.}

\item{factor}{Person-time units for the denominator. Defaults to \code{1000},
e.g., for rates per 1000 person-years of follow-up.}
}
\value{
Tibble:
\itemize{
\item \code{n} Count of participants
\item \code{Events} Count of events
\item \code{Person-time} Sum of person-time
\item \code{Rate per person-time} Incidence rate
\item \code{Risk (Kaplan-Meier, \%)} Cumulative incidence
at the end of follow-up, with log-log 95\% confidence intervals
}
}
\description{
Summary table of overall or stratified incidence rates and cumulative indicence.
To obtain stratified results, group the tibble before calling \code{rates}.
See examples.
}
\examples{
data(ovarian)  # from survival package

# Unstratified
ovarian \%>\%
  rates(time = "futime", event = "fustat")

# Convert time to person-years
ovarian \%>\%
  rates(time = "futime", event = "fustat", conversion = 12)

# Stratified by "ecog.ps":
ovarian \%>\%
  group_by(ecog.ps) \%>\%
  rates(time = "futime", event = "fustat", conversion = 12)
}
\references{
Greenland S, Rothman KJ (2008). Introduction to Categorical Statistics. In:
Rothman KJ, Greenland S, Lash TL. Modern Epidemiology, 3rd edition.
Philadelpha, PA: Lippincott Williams & Wilkins. Page 242. (Poisson/large-sample
approximation for variance of incidence rates)
}
