% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{roundp}
\alias{roundp}
\title{Round and Format p-Values}
\usage{
roundp(pee, equal = FALSE)
}
\arguments{
\item{pee}{Numeric p-vale}

\item{equal}{Show equal sign if p > 0.001? Defaults to \code{FALSE}}
}
\value{
String: rounded number, preceded by equal sign (unless \code{equal = FALSE})
or less-than sign.
}
\description{
Following conventions of many medical journals, this function
rounds a p-value to two digits (if p > 0.05) or three digits (if 0.001 <= p < 0.05),
or returns an inequality of "< 0.001" (if p < 0.001).
}
\examples{
roundp(0.1138)
roundp(c(0.01138, 0.0000138))

# Include equal sign
roundp(0.01138, equal = TRUE)
}
