% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{saveRDS_safely}
\alias{saveRDS_safely}
\title{Safely Save Single Object Without Overwriting the RDS File}
\usage{
saveRDS_safely(..., file = stop("'file' must be specified"))
}
\arguments{
\item{...}{Data frame or other object, and further arguments,
passed on to \code{\link[base]{saveRDS}}. Required.}

\item{file}{Path/file name to for output. Required.}
}
\value{
None.
}
\description{
Wraps \code{\link[base]{saveRDS}}. If the file already exists,
it will not be overwritten. A message will be printed indicating whether
the file was successfully written or if it already existed.
}
\examples{
\dontrun{
data(mtcars)
mtcars \%>\%
  saveRDS_safely(file = "dataset.rds")
}
}
