% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{scale_to_range}
\alias{scale_to_range}
\title{Scale Numeric Vector to a Given Range}
\usage{
scale_to_range(x, min = 0, max = 1, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{min}{Optional. Desired new minimum value. Defaults to 0.}

\item{max}{Optional. Desired new maximum value. Defaults to 1.}

\item{...}{Optional. Passed on to \code{min()} and \code{max()}
functions internally. If vector contains missing values,
may want to specify \code{na.rm = TRUE}.}
}
\value{
Numeric vector with minimum value of \code{min} and
maximum value of \code{max}.
}
\description{
Scale Numeric Vector to a Given Range
}
\examples{
# Simple numeric vector:
x <- c(-2, 0, 4)

# Scale to the c(0, 1) range:
scale_to_range(x)

# Leave minimum value untouched:
scale_to_range(x, min = -2, max = 1)
}
