% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stripplot}
\alias{stripplot}
\title{Box and Whiskers/Jittered Dot Plot}
\usage{
stripplot(
  data,
  x,
  y,
  contrast = NULL,
  unit = NULL,
  digits = 2,
  jitter = TRUE,
  color = NULL,
  na.rm = FALSE,
  printplot = FALSE
)
}
\arguments{
\item{data}{Data frame. Required.}

\item{x}{Categorical variable for x-axis. Required.}

\item{y}{Continuous variable for y-axis. Required.}

\item{contrast}{If added, the mean difference between extreme categories
will be added. \code{contrast} provides the preceding label. See example.
Defaults to \code{NULL} (do not show contrast).}

\item{unit}{Scale to print after the point estimate for the
contrast between extreme categories.
Defaults to \code{NULL} (no unit).}

\item{digits}{Number of digits for rounding point estimate and
confidence intervals of the contrast. Defaults to 2.}

\item{jitter}{Avoid overplotting of points with similar values?
Defaults to \code{TRUE}.}

\item{color}{Variable to color data points by. Defaults to \code{NULL} (none).}

\item{na.rm}{Remove "NA" category from x-axis?
Defaults to \code{FALSE}.}

\item{printplot}{\code{print()} the plot? Defaults to \code{FALSE},
i.e., just return the plot.}
}
\value{
ggplot object, or nothing
(if plot is sent to graphics device with \code{printplot = TRUE}).
Standard customization options for a \code{ggplot} object can be added
on afterwards; see example.
}
\description{
Plots categorical x-axis and continuous y-axis.
Inspired by Nick Cox's Stata plug-in \code{stripplot}. As per
\code{\link[ggplot2]{geom_boxplot}}, boxes reach from the first to the
third quartile; whiskers extend 1.5 times the interquartile range
but not beyond the most extreme data point.
}
\examples{
data(mtcars)

# Basic plot
mtcars \%>\%
  stripplot(x = gear, y = mpg)

# Add mean difference between extreme categories, reduce digits,
# add color by 'wt', add different color scale, and label,
# all using standard ggplot syntax.
mtcars \%>\%
  stripplot(x = gear, y = mpg,
  contrast = "5 vs. 3 gears", unit = "mpg\n", digits = 1, color = wt) +
  scale_color_viridis(option = "cividis") +
  labs(y = "Miles per gallon", color = "Weight")
}
