% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_footnote}
\alias{tab_footnote}
\title{Alternative footnote for gt tables exported as kable}
\usage{
tab_footnote(
  data,
  footnote,
  locations = NULL,
  placement = c("auto", "right", "left")
)
}
\arguments{
\item{data}{gt table or kable table}

\item{footnote}{Footnote. Will be ignored for a kable table.}

\item{locations}{Location, as per gt, e.g., \code{\link[gt]{cells_body}}.
Will be ignored for a kable table.}

\item{placement}{Where to affix footnote marks. Will be ignored for a kable
table.}
}
\value{
gt table or kable table, whichever format \code{data} is in.
}
\description{
This function can be called on a gt table and will the function just as
\code{\link[gt]{tab_footnote}}. If called on a kable table, it will
print the kable table, ignoring the footnote.
}
