% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_style}
\alias{tab_style}
\title{Alternative styling for gt tables exported as kable}
\usage{
tab_style(data, style, locations)
}
\arguments{
\item{data}{gt table or kable table}

\item{style}{Style definition, e.g., per \code{\link[gt]{cell_text}}. Will
be ignored for a kable table.}

\item{locations}{Location, as per gt, e.g., \code{\link[gt]{cells_body}}.
Will be ignored for a kable table.}
}
\value{
gt table or kable table, whichever format \code{data} is in.
}
\description{
This function can be called on a gt table and will function just as
\code{\link[gt]{tab_footnote}}. If called on a kable table, it will
print the kable table, ignoring the styling.
}
