% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table1}
\alias{table1}
\title{Stratified Descriptive Tables Using gtsummary}
\usage{
table1(
  data,
  ...,
  by = NULL,
  overall = TRUE,
  label = "",
  digits = list(all_continuous() ~ c(1, 1)),
  statistic = NULL
)
}
\arguments{
\item{data}{Data frame/tibble to print. Required.}

\item{...}{Optional: Rows to include in table. If none are provided,
all variables will be shown. Supports tidy evaluation; see examples.}

\item{by}{Optional: Stratification variable for columns.}

\item{overall}{Optional: Show "Overall" column if stratifying? Defaults to \code{TRUE}.}

\item{label}{Optional: Label for strata. If empty, the variable label or
variable name of the \code{by} variable will be used.}

\item{digits}{Optional: Level of precision (decimal digits) for statistics.
By default, all continuous variables are shown with one decimal digits
for each statistic: \code{all_continuous() ~ c(1, 1)}.
Precision can also be changed for individual variables, e.g.:
\code{all_continuous() ~ c(1, 1), a_specific_variable ~ c(2, 2)}}

\item{statistic}{Optional: Summary statistic to the shown for specific variables, e.g.:
\code{list(a_specific_variable ~ "{min}, {max}")}.
Passed on to \code{\link[gtsummary]{tbl_summary}(statistic = ...)}.}
}
\value{
Formatted table. Continuous variables are
median (quartile 1, quartile 3); categorical variables are
count (column percent).
}
\description{
\code{table1} wraps \code{\link[gtsummary]{tbl_summary}} with some custom options:
\itemize{
\item Condensed formatting for better space use.
\item Stratification in columns with informative spanning headers, plus an "overall" column if desired.
\item First row with total count.
\item Rounding of continuous values to one decimal digit by default.
}
}
\examples{
data(mtcars)

# Stratify by "gear":
mtcars \%>\%
  table1(by = gear)

# Show only selected variables in rows:
mtcars \%>\%
  table1(cyl, mpg, disp, by = gear)
}
