% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Table 2: Counts and Cox regression results}
\usage{
table2(design, data, layout = "rows", factor = 1000)
}
\arguments{
\item{design}{Data frame that sets up the table. See Details.}

\item{data}{Dataset to be used for all analyses.}

\item{layout}{\code{"rows"} uses the design as rows and
exposure categories as columns. \code{"cols"} is the
opposite: design as columns and exposure categories as rows.}

\item{factor}{Used for \code{type = "rates"}: Factor to multiply
events per person-time by. Defaults to 1000.}
}
\value{
Tibble. Get formatted output by passing on to
\code{\link[khsmisc]{mygt}}.
}
\description{
This function obtains descriptive
and inferential results for survival analyses in the format
of a table stratified by exposure and, if requested, by
effect modifiers.
}
\details{
The main input parameter is the dataset \code{design},
which must consist of the following variables:
\itemize{
\item \code{label} A label for each row (or column).
\item \code{time} The time variable of the survival model.
\item \code{event} The event variable of the survival model.
Events must be \code{1}, censored observations \code{0}.
\item \code{exposure} The exposure variable. Must be categorical
(factor or logical).
\item \code{effect_modifier} An effect modifier variable.
Use \code{NULL} or \code{NA} to leave blank.
\item \code{stratum} A stratum of the effect modifier.
Use \code{NULL} to leave blank. (\code{NA} will evaluate
observations with missing effect modifier.)
\item \code{confounders} A string in the format
\code{"+ confounder1 + confounder2"} that will be substituted into
into \code{coxph(formula = exposure + confounders)}.
Use \code{""} (empty string) to leave blank. Can add \code{"+ strata(site)"}
to obtain Cox models with stratification by, e.g., \code{site}.
\item \code{type} The statistic requested:
\itemize{
\item \code{"cox"} Hazard ratio and 95\% confidence interval.
Models will be fit separately for each stratum of
the \code{effect_modifier}.
\item \code{"cox_joint"} Hazard ratio and 95\% confidence interval.
A 'joint' model for exposure and effect modifier with a
single reference category will be fit.
\item \code{"events"} Event count
\item \code{"time"} Person-time
\item \code{"total"} Number of observations
\item \code{"events/time"} Events slash person-time
\item \code{"events/total"} Events slash number of observations
\item \code{"rate"} Event rate: event count divided by person-time
\item \code{"blank"} An empty line
}
}

Hint: Use \code{\link[tibble]{tibble}}, \code{\link[tibble]{tribble}}, and
\code{\link[dplyr]{mutate}} to construct the \code{design} dataset, especially
variables that are used repeatedly (e.g., \code{time, event}). See examples.
}
\examples{
# Load 'ovarian' dataset from survival package
data(ovarian)

# The exposure (here, 'rx') must be categorical
ovarian <- ovarian \%>\%
  as_tibble() \%>\%
  mutate(rx = factor(rx),
         futime = futime / 365.25)  # presumably years

# Set design elements that vary by row
tribble(
   ~label,                      ~stratum, ~confounders, ~type,
  "Overall: Events",            NULL,     "",           "events",
  "Overall: Person-years",      NULL,     "",           "time",
  "Overall: Rate/1000 years",   NULL,     "",           "rate",
  "  Unadjusted HR (95\% CI)",   NULL,     "",           "cox",
  "  Age-adjusted HR (95\% CI)", NULL,     "+ age",      "cox",
  "",                           NULL,     "",           "blank",
  "Stratified models",          NULL,     "",           "blank",
  "ECOG PS1 (events/N)",        1,        "",           "events/total",
  "  Unadjusted",               1,        "",           "cox",
  "  Age-adjusted",             1,        "+ age",      "cox",
  "ECOG PS2 (events/N)",        2,        "",           "events/total",
  "  Unadjusted",               2,        "",           "cox",
  "  Age-adjusted",             2,        "+ age",      "cox",
  "",                           NULL,     "",           "blank",
  "Joint model, age-adj.",      NULL,     "",           "blank",
  "  ECOG PS1",                 1,        "+ age",      "cox_joint",
  "  ECOG PS2",                 2,        "+ age",      "cox_joint") \%>\%
# Set design elements that are the same for all rows
  mutate(exposure = "rx", event = "fustat", time = "futime",
        effect_modifier = "ecog.ps") \%>\%
# Generate table2
  table2(data = ovarian)

# Use '\%>\% mygt()' to get formatted version (not shown)
}
