% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tabulate_rowcol}
\alias{tabulate_rowcol}
\title{Contigency Table: Tabulate Rows Times Columns}
\usage{
tabulate_rowcol(data, row, col, maxlevels = 30, rowsums = TRUE, colsums = TRUE)
}
\arguments{
\item{data}{Data frame (tibble).}

\item{row}{Row variable.}

\item{col}{Column variable.}

\item{maxlevels}{Optional. Maximum number of levels for rows and columns.
If a row or column variable has more than this number of levels,
for example because it is actually a continuous variable,
then only a warning, not a hard error, will be shown. Defaults to 30.}

\item{rowsums}{Optional. Show row sums? Defaults to TRUE.}

\item{colsums}{Optional. Show column sums? Defaults to TRUE.}
}
\value{
Tibble.
}
\description{
Tabulates \code{row} variable times \code{col} variable.
Supports tidy evaluation for \code{row} and \code{col}; see examples.
}
\examples{
data(mtcars)
mtcars \%>\%
  tabulate_rowcol(gear, carb)

# Swap rows and columns, turn off totals:
mtcars \%>\%
  tabulate_rowcol(row = carb, col = gear,
                  rowsum = FALSE, colsum = FALSE)
}
