% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tsummary}
\alias{tsummary}
\title{Summary of All Numeric Variables in a Data Frame}
\usage{
tsummary(data, ..., by = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{Data frame (tibble).}

\item{...}{Optional. Variables to summarize. If not provided,
all numeric variables will be summarized. Supports tidy evaluation;
see examples.}

\item{by}{Optional. Categorical variable(s) to stratify results by.}

\item{na.rm}{Optional. Drop missing values from summary statatistics?
If set to \code{FALSE}, summary statistics may be missing in the presence of
missing values. Defaults to \code{TRUE}.}
}
\value{
Tibble, possibly grouped, with the following columns:
\itemize{
\item \code{rows} Row count
\item \code{obs} Count of non-missing observations
\item \code{distin} Count of distinct values
\item \code{min} Minimum value
\item \code{q25} 25th percentile
\item \code{median} Median, 50th percentile
\item \code{q75} 75th percentile
\item \code{max} Maximum value
\item \code{mean} Mean
\item \code{sd} Standard deviation
\item \code{sum} Sum of all values
}
}
\description{
Calculate descriptive summary statistics of
all numeric variables in a given dataset. Optionally, this output
can be stratified by one or more categorical variable(s).
}
\examples{
data(mtcars)
mtcars \%>\%
  tsummary()

# Select specific variables and
# remove some summary statistics:
mtcars \%>\%
  tsummary(mpg, cyl, hp, am, gear, carb) \%>\%
  select(-mean, -sd, -sum)

# Stratify by 'gear':
mtcars \%>\%
  tsummary(mpg, hp, carb, by = gear)

# Stratify by 'gear' and 'am':
mtcars \%>\%
  tsummary(mpg, hp, carb, by = c(am, gear))
}
