% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{write_csv_safely}
\alias{write_csv_safely}
\title{Safely Save CSV File Without Overwriting the File}
\usage{
write_csv_safely(..., file = stop("'file' must be specified"))
}
\arguments{
\item{...}{Data frame and other arguments,
passed on to \code{\link[readr]{write_csv}}. Required.}

\item{file}{Path/file name to for output. Required.}
}
\value{
None.
}
\description{
Wraps \code{\link[readr]{write_csv}}. If the file already exists,
it will not be overwritten. A message will be printed indicating whether the file
was successfully written or if it already existed.
}
\examples{
\dontrun{
data(mtcars)
mtcars \%>\%
  dplyr::select(gear, mpg, am) \%>\%
  write_csv_safely(file = "mtcars.csv")
}
}
