% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brickchart.R
\name{brickchart}
\alias{brickchart}
\title{Brick Chart}
\usage{
brickchart(data, outcome, by, group, colors = NULL, guide = FALSE)
}
\arguments{
\item{data}{Data set.}

\item{outcome}{Outcome expression, e.g., \code{event == TRUE}.}

\item{by}{Exposure variable.}

\item{group}{Optional: Grouping variable, e.g., an effect modifier.}

\item{colors}{Optional: Color list. Must be a \code{list} consisting of
two-element color code vectors with the dark and bright colors for each
level of the exposure variable (\code{by}).
Example: \code{list(c("darkred", "red"), c("darkblue", "lightblue"))}.
If not provided, colors will be generated from the
\code{\link[viridis]{viridis_pal}} palette.}

\item{guide}{Optional: Show legend? Defaults to \code{FALSE}.}
}
\value{
ggplot. Modify further with standard ggplot functions.
}
\description{
Plot bars of proportions that consist of "bricks" showing individual
observations.
}
\examples{
data(cancer, package = "survival")
cancer <- cancer \%>\%
  tibble::as_tibble() \%>\%
  dplyr::mutate(sex = factor(sex, levels = 1:2,
                             labels = c("Men", "Women")))

cancer \%>\%
  dplyr::filter(ph.ecog < 3) \%>\%  # drop missing/near-empty categories
  brickchart(outcome = status == 2,
             by = ph.ecog)

# Stratified version
# Note- Color fill may be off with ggplot v3.3.4+
cancer \%>\%
  dplyr::filter(ph.ecog < 3) \%>\%
  brickchart(outcome = status == 2,
             by = ph.ecog,
             group = sex,
             guide = TRUE) +  # show color legend
  # Modify graph with standard ggplot functions
  # Refer to axes before flipping x <-> y. Here, y is horizontal:
  ggplot2::labs(y = "Risk (cumulative incidence)",
                fill = "ECOG status",  # Color label
                title = "Mortality by ECOG performance status") +
  # Themes refer to axes as shown--'y' is now vertical
  ggplot2::theme(axis.title.y = ggplot2::element_blank())

}
