% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrmat}
\alias{corrmat}
\title{Correlation plot}
\usage{
corrmat(
  data,
  ...,
  method = "pearson",
  use = "pairwise.complete.obs",
  reorder = TRUE,
  digits = 2,
  fontsize = 4,
  legendpos = c(0.15, 0.35),
  cutpoints = c(-1, 0, 1),
  colors = c((viridis::viridis_pal(option = "cividis"))(10)[1], "white",
    (viridis::viridis_pal(option = "cividis"))(10)[8])
)
}
\arguments{
\item{data}{Required. Data frame. Only numeric variables will be used;
variables of other types will silently be dropped. To include
categorical variables, coerce factors to numeric first. See examples.}

\item{...}{Optional. Variables to correlate. If not provided,
all numeric variables will be used. Supports tidy evaluation;
see examples.}

\item{method}{Optional. Correlation method. Defaults to \code{"pearson"}.}

\item{use}{How \code{\link[stats]{cor}} handles
missing values. Defaults to a restriction to complete
observations with all variables. See parameter \code{use} of
\code{\link[stats]{cor}} for alternatives.}

\item{reorder}{Perform hierarchical clustering in the correlation matrix?
This will order variables by their correlation patterns with other
variables. If turned off to \code{FALSE}, the order of variables in the
dataset will be retained. Defaults to \code{TRUE}.}

\item{digits}{Decimal digits for displayed correlation coefficients.
Defaults to \code{2}.}

\item{fontsize}{Size for text in boxes and to the left of the boxes ("y
axis"). Defaults to \code{4}.}

\item{legendpos}{(x, y) coordinates of color legend. Use
\code{legendpos = "none"} to turn off the legend.
Defaults to \code{c(0.15, 0.35)}.}

\item{cutpoints}{Correlation coefficient values that have a distinct
color. Defaults to \code{c(-1, 0, 1)}.}

\item{colors}{Colors for the \code{cutpoints}. Defaults to blue (for
negative), white (no correlation), and yellow (positive correlation) on the
\code{cividis} color scale.}
}
\value{
ggplot. Can be modified with the usual ggplot commands, such as
\code{\link[ggplot2]{theme}}.
}
\description{
Performs pairwise Pearson or Spearman correlations.
}
\examples{
data(mtcars)

mtcars \%>\%
  corrmat(mpg, cyl, hp, wt, qsec)

# Can use tidy evaluation to select variables:
mtcars \%>\%
  corrmat(contains("a"), starts_with("c"))

# If "cyl" was a character, it would be excluded:
mtcars \%>\%
  dplyr::mutate(cyl_chr = as.character(cyl)) \%>\%
  corrmat(mpg, cyl_chr, hp, wt, qsec)

# To retain the character variable "cyl",
# convert to factor and then make numeric:
mtcars \%>\%
  dplyr::mutate(cyl_chr = as.character(cyl),
                cyl_chr = as.numeric(factor(cyl_chr))) \%>\%
  corrmat(mpg, cyl_chr, hp, wt, qsec)
}
\seealso{
\url{http://www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization}
}
