% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mygt}
\alias{mygt}
\title{Turn tibble into gt Table with Custom Formatting}
\usage{
mygt(df, md = NULL)
}
\arguments{
\item{df}{Data frame/tibble}

\item{md}{Optional. If not \code{NULL}, then the given
columns will be printed with markdown formatting, e.g., \code{md = c(1, 3)}
for columns 1 and 3.}
}
\value{
Formatted gt table
}
\description{
Formatting includes:
\itemize{
\item Text align to top/left
\item Smaller row padding
\item No top border
\item Bold column labels
}
}
\examples{
data(mtcars)
mtcars \%>\%
  dplyr::slice(1:5) \%>\%
  mygt()
}
