% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rates}
\alias{rates}
\title{Event Counts, Person-Time, Incidence Rates, and Cumulative Incidence (Risk).}
\usage{
rates(
  data,
  time,
  event,
  by,
  risk_time = NULL,
  conversion = 1,
  rate_digits = 1,
  risk_digits = 1,
  factor = 1000
)
}
\arguments{
\item{data}{Data frame/tibble. If grouped using \code{\link[dplyr]{group_by}},
then stratified results will be returned.}

\item{time}{Variable indicating follow-up time.}

\item{event}{Variable indicating events (event = 1, censor = 0).}

\item{by}{Optional. Variable to stratify results by. Defaults
to how the dataset is grouped, i.e., if it is grouped already,
results will be shown by the stratification variable.}

\item{risk_time}{Optional. Time point at which cumulative incidence
(risk) should be calculated. Provide in original units of follow-up
time, i.e., before potential division by the \code{conversion} factor.
If not provided, defaults to the end of follow-up, which may not
be the best interpretable timepoint.}

\item{conversion}{Optional. Divisor for follow-up time. Defaults to 1.
Set \code{conversion = 12} if \code{time} is in months and
rates per person-years are desired.}

\item{rate_digits}{Optional. Decimal digits to display for rates. Defaults to 1.}

\item{risk_digits}{Optional. Decimal digits to display for
cumulative incidence (risk). Defaults to 1.}

\item{factor}{Optional. Person-time units for the denominator. Defaults to \code{1000},
e.g., for rates per 1000 person-years of follow-up.}
}
\value{
Tibble:
\itemize{
\item \code{n} Count of participants.
\item \code{Events} Count of events.
\item \code{Person-time} Sum of person-time.
\item \code{Rate per person-time} Incidence rate.
\item \code{Risk (\%)} Cumulative incidence.
at time \code{risk_time} or, if not provided, at the end of follow-up,
with log-log 95\% confidence intervals.
}
}
\description{
Summary table of overall or stratified incidence rates and cumulative indicence.
To obtain stratified results, group the tibble before calling \code{rates}.
See examples.
}
\examples{
data(ovarian)  # from survival package

# Unstratified
ovarian \%>\%
  rates(time = futime, event = fustat)

# Show risk (cumulative incidence) at 25 years,
# show follow-up time to person-years
ovarian \%>\%
  rates(time = futime, event = fustat,
        risk_time = 25 * 12, conversion = 12)

# Stratified by "ecog.ps" and "resid.ds":
ovarian \%>\%
  rates(time = futime, event = fustat,
        by = c(ecog.ps, resid.ds),
        risk_time = 25 * 12, conversion = 12)
}
\references{
Greenland S, Rothman KJ (2008). Introduction to Categorical Statistics. In:
Rothman KJ, Greenland S, Lash TL. Modern Epidemiology, 3rd edition.
Philadelpha, PA: Lippincott Williams & Wilkins. Page 242. (Poisson/large-sample
approximation for variance of incidence rates)
}
